[7419] Fixed some type convertion warnings.

+++ src/game/SpellEffects.cpp
@@ -1255,7 +1255,7 @@ void Spell::EffectDummy(uint32 i)
             if (m_spellInfo->SpellFamilyFlags & 0x0000000000040000LL)
             {
                 // In 303 exist spirit depend
-                uint32 spirit = m_caster->GetStat(STAT_SPIRIT);
+                uint32 spirit = uint32(m_caster->GetStat(STAT_SPIRIT));
                 switch (m_spellInfo->Id)
                 {
                     case  1454: damage+=spirit; break;
